buffered gene_tables inserts. Works now

+++ gemini/droptables.py
@@ -2,8 +2,12 @@
 
 from cassandra.cluster import Cluster
 
-cluster = Cluster()
-session = cluster.connect()
-session.execute('DROP KEYSPACE IF EXISTS gemini_keyspace')
+answer = raw_input('Drop all gemini tables (y | n)? ')
+
+if answer in ['yes', 'yes please', 'ja', 'ja, graag']:
+	print 'Dropping gemini tables...'
+	cluster = Cluster()
+	session = cluster.connect()
+	session.execute('DROP KEYSPACE IF EXISTS gemini_keyspace')
